requirements.txt
streamlit==1.45.1
openai==1.84.0
supabase==2.15.2
python-dotenv==1.1.0
google-generativeai==0.8.5